[AI] sysfs store and show functions working, all need more testing

+++ AI_gov_types.h
@@ -17,7 +17,7 @@ struct AI_gov_profile{
 	struct kobject*		kobj;
 
 	unsigned int desired_frame_rate;
-	float current_frame_rate;
+	unsigned int current_frame_rate;
 };
 
 struct AI_gov_info{
